+++ core/api/typescript/src/lib/bulk_command_execution/v0_bulk_command_api/v0_bulk_commands.ts
@@ -1,4 +1,15 @@
-import { ExecCommandArgs, ExecuteBulkCommandsArgs, ExecuteModuleArgs, LoadModuleArgs, RegisterServiceArgs, RemoveServiceArgs, RepartitionArgs, StartServiceArgs, WaitForHttpGetEndpointAvailabilityArgs, WaitForHttpPostEndpointAvailabilityArgs } from "../../../kurtosis_core_rpc_api_bindings/node/api_container_service_pb";
+import {
+    ExecCommandArgs,
+    ExecuteBulkCommandsArgs,
+    ExecuteModuleArgs,
+    LoadModuleArgs,
+    RegisterServiceArgs,
+    RemoveServiceArgs,
+    RepartitionArgs,
+    StartServiceArgs,
+    WaitForHttpGetEndpointAvailabilityArgs,
+    WaitForHttpPostEndpointAvailabilityArgs
+} from "../../../";
 import { V0CommandTypeVisitor, V0CommandType } from "./v0_command_types";
 import { ok, err, Result } from "neverthrow";
 import * as protobuf from "google-protobuf";