+++ core/api/typescript/src/lib/constructor_calls.ts
@@ -31,7 +31,7 @@ import * as jspb from "google-protobuf";
 // ==============================================================================================
 //                           Shared Objects (Used By Multiple Endpoints)
 // ==============================================================================================
-export function newPort(number: number, protocol: Port.ProtocolMap[keyof Port.ProtocolMap]) {
+export function newPort(number: number, protocol: Port.Protocol) {
     const result: Port = new Port();
     result.setNumber(number);
     result.setProtocol(protocol);