+++ core/docs/changelog.md
@@ -2,6 +2,8 @@
 
 ### Features
 * Added Envoy Proxy to support gRPC-web
+* Added interfaces to Service, Module and Enclave Contexts
+
 # 1.38.0
 ### Breaking Changes
 * Change the `ApiContainerLauncher.LaunchWithDefaultVersion()` and `ApiContainerLauncher.LaunchWithCustomVersion()` methods API, adding two new arguments `grpcListenPort` and `grpcProxyListenPort` and deleting the one named `listenPort`